Use a single 'useRowSelection' hook
