+++ src/hooks/index.ts
@@ -6,4 +6,3 @@ export * from './useViewportRows';
 export * from './useFocusRef';
 export * from './useLatestFunc';
 export * from './useRowSelection';
-export * from './useRowSelectionChange';
