+++ src/index.ts
@@ -6,6 +6,7 @@ export * from './Columns';
 export * from './formatters';
 export { default as TextEditor } from './editors/TextEditor';
 export { default as SortableHeaderCell } from './headerCells/SortableHeaderCell';
+export { useRowSelection } from './hooks';
 export type {
   Column,
   CalculatedColumn,
